+++ src/main/java/cz/upce/fei/bookdb_backend/dto/BookResponseDtoV1.java
@@ -38,5 +38,5 @@ public class BookResponseDtoV1 {
     //
     private long reviewCount;
 
-    private Long rating;
+    private Float rating;
 }
